feat: PyCap wrapper now loads params from config
Due to the number of steps that can be run at once and the sheer
number of parameters, it would be unwieldy to expect users to
use command-line flags for all parameters.

Now, params are specified in a config .yml file which may be
slightly more complicated at first but will make the process
much easier. see examples/config.py a for an example.